--- a/src/decimal.rs
+++ b/src/decimal.rs
@@ -208,6 +208,7 @@ impl Decimal {
     /// let num = Decimal::new(1234, 3);
     /// assert_eq!(num.scale(), 3u32);
     /// ```
+    #[inline]
     pub fn scale(&self) -> u32 {
         ((self.flags & SCALE_MASK) >> SCALE_SHIFT) as u32
     }
@@ -792,7 +793,7 @@ fn rescale(left: &mut [u32; 3], left_scale: &mut u32, right: &mut [u32; 3], righ
     };
 
     let mut working = [my[0], my[1], my[2]];
-    while diff > 0 && mul_by_u32(&mut working, 10) == 0 {
+    while diff > 0 && mul_by_10(&mut working) == 0 {
         my.copy_from_slice(&working);
         diff -= 1;
     }
@@ -1118,6 +1119,22 @@ fn sub_part(left: u32, right: u32, overflow: u32) -> (u32, u32) {
     }
     (lo, hi as u32)
 }
+
+// Returns overflow
+#[inline]
+fn mul_by_10(bits: &mut [u32; 3]) -> u32 {
+    let mut overflow = 0u64;
+    for b in bits.iter_mut() {
+        let result = u64::from(*b) * 10u64 + overflow;
+        let hi = (result >> 32) & U32_MASK;
+        let lo = (result & U32_MASK) as u32;
+        *b = lo;
+        overflow = hi;
+    }
+
+    overflow as u32
+}
+
 
 // Returns overflow
 fn mul_by_u32(bits: &mut [u32], m: u32) -> u32 {