This changes gives improvements up to 22% in basic benchmarks

before
```
test add_negative_pi         ... bench:         169 ns/iter (+/- 2)
test add_negative_point_five ... bench:          28 ns/iter (+/- 0)
test add_one                 ... bench:          32 ns/iter (+/- 0)
test add_one_hundred         ... bench:          32 ns/iter (+/- 0)
test add_pi                  ... bench:         167 ns/iter (+/- 5)
test add_point_zero_one      ... bench:          18 ns/iter (+/- 0)
test add_two                 ... bench:          32 ns/iter (+/- 0)
```
after
```
test add_negative_pi         ... bench:         169 ns/iter (+/- 1)
test add_negative_point_five ... bench:          28 ns/iter (+/- 0)
test add_one                 ... bench:          28 ns/iter (+/- 0)
test add_one_hundred         ... bench:          28 ns/iter (+/- 1)
test add_pi                  ... bench:         163 ns/iter (+/- 2)
test add_point_zero_one      ... bench:          14 ns/iter (+/- 1)
test add_two                 ... bench:          28 ns/iter (+/- 0)
```

and huge win in added in PR benchmark 

before
```
test bench_sum_10m           ... bench: 150,744,073 ns/iter (+/- 2,652,292) 
```
after
```
test bench_sum_10m           ... bench:  78,743,056 ns/iter (+/- 1,843,435)
```
The same benchmark with sum of `decimal`s in C#\.net shows perf around 115 ms/iter